feat: Android AAR build support (#2)
* feat: Add Android Gradle wrapper and initial project structure

- Added gradle-wrapper.properties to configure Gradle version 8.11.1.
- Introduced gradlew and gradlew.bat scripts for Unix and Windows environments.
- Created proguard-rules.pro for ProGuard configuration.
- Set up settings.gradle.kts for dependency management.
- Added AndroidManifest.xml with permissions for external storage access.
- Implemented PoDoFoException class for custom exception handling.
- Developed PoDoFoWrapper class as a Java wrapper for the native PoDoFo library.
- Created CMakeLists.txt for building the native library.
- Added JNI header and implementation files for interfacing with the native code.

* feat: Add Android Gradle wrapper and initial project structure

- Added gradle-wrapper.properties to configure Gradle version 8.11.1.
- Introduced gradlew and gradlew.bat scripts for Unix and Windows environments.
- Created proguard-rules.pro for ProGuard configuration.
- Set up settings.gradle.kts for dependency management.
- Added AndroidManifest.xml with permissions for external storage access.
- Implemented PoDoFoException class for custom exception handling.
- Developed PoDoFoWrapper class as a Java wrapper for the native PoDoFo library.
- Created CMakeLists.txt for building the native library.
- Added JNI header and implementation files for interfacing with the native code.

* fix: update project group and version for Android support

* feat: Add support for additional X.509 certificates in CMS signing

- Introduced `AddCertificate` method in `PdfSignerCms` to allow adding extra certificates to the CMS "certificates" collection, enabling the embedding of intermediate/chain certificates.
- Updated `resetContext` to reapply added certificates after context reset.
- Implemented `AddCertificate` in `CmsContext` to handle the addition of certificates, supporting both DER and PEM formats.
- Enhanced `PdfRemoteSignDocumentSession` to accept a list of chain certificates during initialization.
- Added JNI wrapper methods to facilitate the use of additional certificates from Java.

* fix: remove push trigger for build-documentation workflow

* refactor: move OpenSSL BIO management to implementation file and update header to remove private dependencies

* feat: add GitHub Actions workflow for building Android AAR

* fix: update NDK_VERSION to r27c for Android build workflow

* feat: update Android build configuration and dependencies for improved compatibility

* feat: integrate ccache support for improved build performance in Android scripts

+++ src/podofo/private/XmlUtils.cpp
@@ -196,7 +196,7 @@ void utls::NavigateDescendantElements(xmlNodePtr element, const string_view& ns,
 
 bool utls::TrySerializeXmlDocTo(string& str, xmlDocPtr doc)
 {
-    auto ctx = xmlSaveToIO(xmlOutputStringWriter, xmlOutputStringWriterClose, &str, nullptr, XML_SAVE_NO_DECL | XML_SAVE_FORMAT);
+    auto ctx = xmlSaveToIO(xmlOutputStringWriter, xmlOutputStringWriterClose, &str, nullptr, (int)(XML_SAVE_NO_DECL | XML_SAVE_FORMAT));
     if (ctx == nullptr || xmlSaveDoc(ctx, doc) == -1 || xmlSaveClose(ctx) == -1)
         return false;
